Start sending error and received emails for form submissions
